added background and overlay, styling
