+++ webapp/src/App.js
@@ -11,9 +11,9 @@ import Register from "./components/Register";
 import Simulation from "./components/Simulation";
 import Candle from "./components/Candle";
 import Portfolio from "./components/Portfolio";
-import Search from "./components/Search";
-import Friends from "./components/Friends";
-import FriendRequests from "./components/FriendRequests";
+import Search from "./components/social/Search";
+import Friends from "./components/social/Friends";
+import FriendRequests from "./components/social/FriendRequests";
 import Menu from "./components/Menu";
 import bull from "./bull.png";
 import br from "./triangles.svg";
@@ -63,36 +63,41 @@ export default function App() {
   return (
     <>
       <Nav />
-      <Routes>
-        <Route path="/" element={<HomePage />} />
-        <Route path="/about" element={<About />} />
-        <Route path="/blog" element={<Blog />} />
-        <Route path="/post" element={<Post />} />
-        <Route path="/dashboard" element={<Dashboard />} />
-        <Route
-          path="/login"
-          element={<Login loggedIn={loggedIn} setLoggedIn={setLoggedIn} />}
-        />
-        <Route path="/logout" element={<Logout setLoggedIn={setLoggedIn} />} />
-        <Route path="/register" element={<Register />} />
-        <Route
-          path="/simulation"
-          element={<Simulation loggedIn={loggedIn} />}
-        />
-        <Route path="/candle" element={<Candle />} />
-        <Route
-          path="/portfolio"
-          element={
-            <Portfolio
-              username={localStorage.getItem("username")}
-              isSelf={true}
-            />
-          }
-        />
-        <Route path="/search" element={<Search />} />
-        <Route path="/friends" element={<Friends />} />
-        <Route path="/friend-requests" element={<FriendRequests />} />
-      </Routes>
+      <div className="overlay">
+        <Routes>
+          <Route path="/" element={<HomePage />} />
+          <Route path="/about" element={<About />} />
+          <Route path="/blog" element={<Blog />} />
+          <Route path="/post" element={<Post />} />
+          <Route path="/dashboard" element={<Dashboard />} />
+          <Route
+            path="/login"
+            element={<Login loggedIn={loggedIn} setLoggedIn={setLoggedIn} />}
+          />
+          <Route
+            path="/logout"
+            element={<Logout setLoggedIn={setLoggedIn} />}
+          />
+          <Route path="/register" element={<Register />} />
+          <Route
+            path="/simulation"
+            element={<Simulation loggedIn={loggedIn} />}
+          />
+          <Route path="/candle" element={<Candle />} />
+          <Route
+            path="/portfolio"
+            element={
+              <Portfolio
+                username={localStorage.getItem("username")}
+                isSelf={true}
+              />
+            }
+          />
+          <Route path="/search" element={<Search />} />
+          <Route path="/friends" element={<Friends />} />
+          <Route path="/friend-requests" element={<FriendRequests />} />
+        </Routes>
+      </div>
     </>
   );
 }
