+++ webapp/src/app.css
@@ -21,21 +21,32 @@ html * {
   /* Preserve aspet ratio */
   min-width: 100%;
   min-height: 100%;
-  z-index: -10;
+}
+
+.overlay {
+  padding-left: 1vw;
+  padding-right: 1vw;
+  background-color: rgba(0, 0, 0, 0.5);
+  height: 100%;
+  width: 80%;
+  position: fixed;
+  top: 56vh;
+  left: 50%;
+  transform: translate(-50%, -50%);
 }
 
 .home {
+  padding: 1em;
   background-color: rgb(74, 163, 248);
-  position: absolute;
+  position: relative;
+  top: 50%;
+  left: 50%;
+  max-width: 10em;
+  transform: translate(-50%, -50%);
   display: flex;
-  height: 35vh;
-  width: 40vh;
   flex-direction: column;
   justify-content: space-around;
   align-items: center;
-  top: 40vh;
-  left: 50vw;
-  transform: translate(-50%, -50%);
   font-size: 3vh;
   border-radius: 2vh;
   text-align: center;
@@ -60,6 +71,8 @@ html * {
 }
 
 .nav {
+  position: fixed;
+  width: 100%;
   height: 6vh;
   display: flex;
   flex-direction: row;
@@ -130,5 +143,9 @@ a:active {
 
   .menu * {
     display: block;
+  }
+
+  .overlay {
+    width: 100%;
   }
 }
