+++ webapp/src/components/social/search.css
@@ -1,0 +1,37 @@
+.search {
+  position: relative;
+  left: 50%;
+  font-size: 2.5vh;
+  transform: translate(-50%, 50%);
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+.search h3 {
+  padding-bottom: 2rem;
+}
+
+.search input {
+  font-size: 2.5vh;
+}
+
+.search .user-display {
+  margin-top: 1rem;
+}
+
+.social-tabs {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  padding-bottom: 2rem;
+}
+
+.social-tabs .separator {
+  margin-left: 0.5rem;
+  margin-right: 0.5rem;
+}
+
+.search input {
+  color: black;
+}
