+++ webapp/src/components/style/menu.css
@@ -1,4 +1,5 @@
 .menubutton {
+  background-color: transparent;
   border: none;
   padding-right: 1vh;
 }
@@ -10,7 +11,6 @@
   position: absolute;
   top: 6.5vh;
   left: 63vw;
-  z-index: 12;
   border-radius: 1vh;
 }
