--- webapp/src/components/style/search.css
@@ -1,25 +1,0 @@
-.search {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    text-align: center;
-}
-
-.search h3 {
-    margin-bottom: 1rem;
-}
-
-.search .search-bar {
-    margin-top: 1rem;
-    width: 15%;
-    background-color: aqua;
-    border-radius: 25px;
-    color: black;
-    padding-left: 0.5rem;
-    border-style:double;
-    border-color: coral;
-}
-
-.search .user-display {
-    margin-top: 1rem;
-}
