+++ webapp/src/components/style/simulation.css
@@ -1,6 +1,6 @@
 .simulation {
   background-color: white;
-  position: absolute;
+  position: relative;
   width: min(80%, 700px);
   top: 50%;
   left: 50%;
