--- webapp/src/components/style/social-tabs.css
@@ -1,10 +1,0 @@
-.social-tabs {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-}
-
-.social-tabs .separator {
-    margin-left: 0.5rem;
-    margin-right: 0.5rem;
-}
